Moving sTGC RDO charge to uint16_t from float 	modified:   MuonSpectrometer/MuonCnv/MuonByteStreamCnvTest/src/STGC_DigitToRDO.cxx 	modified:   MuonSpectrometer/MuonRDO/MuonRDO/STGC_RawData.h 	modified:   MuonSpectrometer/MuonRDO/src/STGC_RawData.cxx 	modified:   MuonSpectrometer/MuonValidation/MuonPRDTest/src/sTGCRDOVariables.cxx 	modified:   MuonSpectrometer/MuonValidation/MuonPRDTest/src/sTGCRDOVariables.h
Former-commit-id: 4f96f8f52b93c3b70bb80667fc91930f56e770b5

+++ MuonSpectrometer/MuonValidation/MuonPRDTest/src/sTGCRDOVariables.h
@@ -65,7 +65,7 @@ class sTGCRDOVariables : public ValAlgVariables
   std::vector<int> *m_NSWsTGC_rdo_channel;
   std::vector<int> *m_NSWsTGC_rdo_channel_type;
   std::vector<double> *m_NSWsTGC_rdo_time;
-  std::vector<double> *m_NSWsTGC_rdo_charge;
+  std::vector<uint16_t> *m_NSWsTGC_rdo_charge;
   std::vector<uint16_t> *m_NSWsTGC_rdo_bcTag;
   std::vector<bool> *m_NSWsTGC_rdo_isDead;
